Provided configurable CFPClient via project devoxx-cfp-impl and TestBase to test the configured client via project devoxx-cfp-test. Changed the existing CFPClients and their tests to use the configurable CFPClient fixes TweetWallFX/TweetwallFX#330
