--- devoxx-2017-be/cfp-impl/src/main/resources/META-INF/services/org.tweetwall.devoxx.api.cfp.client.CFPClient
@@ -1,1 +1,0 @@
-org.tweetwall.devoxx.api.cfp.client.impl2017be.CFPClientDevoxx2017BE
