+++ devoxx-2017-be/cfp-impl/src/main/resources/tweetwallConfig.json
@@ -1,0 +1,7 @@
+{
+    "devoxxCFP": {
+        "baseUri": "https://cfp.devoxx.be/api",
+        "eventId": "DVBE17",
+        "votingResultsUri": "https://cfp.devoxx.be/api/voting/v1/top/talks"
+    }
+}
