+++ devoxx-2018-pl/cfp-impl/build.gradle
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright 2014-2017 TweetWallFX
+ * Copyright 2014-2018 TweetWallFX
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,8 +23,10 @@
  */
 
 dependencies {
-    compile project(':TweetWallFX-Devoxx-API-CFP')
     compile group: 'org.apache.logging.log4j', name: 'log4j-api', version: versionLog4j
 
+    runtime project(':TweetWallFX-Devoxx-CFP-Impl')
     runtime group: 'org.apache.logging.log4j', name: 'log4j-core', version: versionLog4j
+
+    testCompile project(':TweetWallFX-Devoxx-CFP-Test')
 }
