--- devoxx-2018-pl/cfp-impl/src/main/resources/META-INF/services/org.tweetwall.devoxx.api.cfp.client.CFPClient
@@ -1,1 +1,0 @@
-org.tweetwall.devoxx.api.cfp.client.impl2018pl.CFPClientDevoxx2018PL
