--- devoxx-api-cfp/src/test/resources/META-INF/services/org.tweetwall.devoxx.api.cfp.client.CFPClient
@@ -1,1 +1,0 @@
-org.tweetwall.devoxx.api.cfp.client.CFPClientTestImpl
