+++ devoxx-cfp-impl/src/main/resources/META-INF/services/org.tweetwall.devoxx.api.cfp.client.CFPClient
@@ -1,0 +1,1 @@
+org.tweetwall.devoxx.cfp.impl.ConfigurableCFPClientImpl
