+++ devoxx-cfp-impl/src/main/resources/META-INF/services/org.tweetwallfx.config.ConfigurationConverter
@@ -1,0 +1,1 @@
+org.tweetwall.devoxx.cfp.impl.CFPClientSettings$Converter
