+++ devoxx-cfp-impl/src/test/resources/log4j2.xml
