+++ settings.gradle
@@ -47,6 +47,10 @@ include ':devoxx-2015'
 project(":devoxx-2015").name = "TweetWallFX-Devoxx-2015"
 include ':devoxx-api-cfp'
 project(":devoxx-api-cfp").name = "TweetWallFX-Devoxx-API-CFP"
+include ':devoxx-cfp-impl'
+project(":devoxx-cfp-impl").name = "TweetWallFX-Devoxx-CFP-Impl"
+include ':devoxx-cfp-test'
+project(":devoxx-cfp-test").name = "TweetWallFX-Devoxx-CFP-Test"
 include ':devoxx-2016'
 project(":devoxx-2016").name = "TweetWallFX-Devoxx-2016"
 include ':devoxx-2017-fr'
