--- voxxed-days-zurich/src/main/resources/META-INF/services/org.tweetwall.devoxx.api.cfp.client.CFPClient
@@ -1,1 +1,0 @@
-org.tweetwallfx.vdz.cfp.CFPClientVDZ18
